requirements.txt
# python 3.10
# for GCN, GAT, SIMPLE-HGN, Heco, TGN, DDGCL, MDGPT, MDGFM, SAMGPT, GCOPE, GraphCLIP, G2P2

torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
-f https://download.pytorch.org/whl/cu121

#py-tgb>=2.0.0
#ipykernel
pykeen>=1.11.1
ogb>=1.3.6
rdkit

torch_geometric
pyg_lib
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
-f https://data.pyg.org/whl/torch-2.5.0+cu121.html

# GCOPE
torchmetrics==1.2.0
terminaltables==3.1.10
-f https://github.com/Phoveran/fastargs.git@main#egg=fastargs

# GraphCLIP
vllm==0.7.0  #==0.7.3
transformers==4.47.0

# G2P2
ftfy

-f https://data.dgl.ai/wheels/torch-2.3/cu121/repo.html
dgl==2.5.0

# DGI
pygcl>=0.1.2
matplotlib

# Unigraph2
pytorch_lightning>=2.4.0
wandb>=0.22.2